Remove godot_header environment variable check
This was a left over from when this module was in a separate folder. Now that we have godot_headers as a submodule there is no need to override this in this way. Its even questionable if we need to check for a command line override.  The changes of the environment variable pointing to an old version has proven to be very likely..