Fix occurrence of infinity transform from pose.
OpenXR may give a zero length quaternion at initialization causing a
transform with inf values, resulting in a slew of assertion errors from
`VisualServer.instance_set_transform` if a VisualInstance happens to be
a child of OpenXRHand, OpenXRPose or OpenXRSkeleton.

The OpenXR specification states that the XrQuaterionf should be of unit
length, so I figured `is_normalized` is a good health check.